feat: Add platform-specific content storage implementations for Linux and macOS
